# Include evaluation performance in performance summary