1. Additional counters for project evaluation.
2. Update ConfigurableForwardingLogger to forward BuildStatusEvents
a. This ensures that project evaluation started/finished events are piped through in multi-proc builds so the performance summary can correctly contain evaluation numbers.  The events are only forwarded if the verbosity is diagnostic or the performance summary will be shown.

Summary looks like this:

```
Project Evaluation Performance Summary:
       25 ms  D:\CommonBuildToolset\Examples\src\CSharp\Windows\dirs.proj   1 calls
       29 ms  D:\CommonBuildToolset\Examples\src\CSharp\Windows\UnitTests.MSTest\UnitTests.MSTest.csproj   1 calls
       30 ms  D:\CommonBuildToolset\Examples\src\CSharp\Windows\UnitTests.NUnit\UnitTests.NUnit.csproj   1 calls
       74 ms  D:\CommonBuildToolset\Examples\src\CSharp\Windows\ClassLibrary\ClassLibrary.csproj   2 calls
      138 ms  D:\CommonBuildToolset\Examples\src\CSharp\Windows\ClassLibrary.Package\ClassLibrary.Package.nuproj   1 calls
      220 ms  D:\CommonBuildToolset\Examples\src\dirs.proj   1 calls
```